Add debug logger; Add ws upgrader options; Move handlers to separate file;
 Added logger support to server, standard logger can de used as in examples. Added upgrader to server options to set parameters for web socket connection.
